+++ dotnet/src/RepositoryContract/LinqSupport/RepositoryExpressionExtensions.Aggregation.cs
@@ -4,7 +4,6 @@ using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Data.Fuse.Convenience;
 using System.Linq;
-using System.Linq.Dynamic.Core;
 using System.Linq.Expressions;
 using System.Reflection;
 using System.Security.Principal;