+++ dotnet/src/RepositoryContract/LinqSupport/RepositoryExpressionExtensions.cs
@@ -3,7 +3,6 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Data.Fuse.Convenience;
 using System.Linq;
-using System.Linq.Dynamic.Core;
 using System.Linq.Expressions;
 using System.Reflection;
 using System.Security.Principal;